+++ package.json
@@ -19,7 +19,8 @@
   "homepage": "https://github.com/LingDong-/wenyan-lang",
   "scripts": {
     "build": "npm run clear && webpack --mode production",
-    "build:dev": "npm run clear && webpack --mode development",
+    "build:site": "webpack --mode development --config webpack.site.config.js",
+    "dev:site": "webpack --mode development --config webpack.site.config.js --watch",
     "dev": "npm run clear && webpack --mode development --watch",
     "clear": "rimraf ./dist",
     "publish": "bump --commit && node ./tools/publish.js",
@@ -27,10 +28,6 @@
     "release": "bump --commit --tag && git push --follow-tags",
     "docs:update": "node ./tools/stdlib_doc.js",
     "make_examples_readme": "node ./tools/make_examples_readme.js",
-    "make_ide": "node ./tools/make_ide.js",
-    "make_site": "node ./tools/make_site.js",
-    "ide:dev": "nodemon --ignore build,dist ./tools/make_ide.js",
-    "site:dev": "nodemon --ignore build,dist ./tools/make_site.js",
     "test": "mocha --require mocha-snapshots",
     "test:update": "mocha --require mocha-snapshots --update",
     "lint:fix": "npm run lint -- --fix",
