--- site/CNAME
@@ -1,1 +1,0 @@
-wenyan-lang.lingdong.works
